Make node_t type smaller for our specific use case.
This optimization is just applicable in our use case, where we have much
fewer nodes than what even fits in a 16-bit integer. Which means we
could maybe cut the size even more if we played a bit with the numbers.
But it's still difficult.

+++ node.h
@@ -8,7 +8,7 @@
 
 extern int node_count;
 
-typedef int64_t node_t;
+typedef int32_t node_t;
 
 /**********************************************************************
  * WARNING: THIS FUNCTION ALLOCATES MEMORY
